Create command server bootstrap script
Creates a shell script to be downloaded and run on the command server
on first boot. This script initially configures the command server and
performs the API calls for the other basic service servers. This script
reads in bootstrap_variables, created by the initialization script run
on a Mac (not yet written).

As of version 0.1, bootstrapping is done via shell script for
simplicity's sake (quick-and-dirty, easy to write). In future versions,
it may be better to use Ansible for initial "bootstrapping" and the
cloud-config "runcmd" to install Ansible.