Create Lotus initialization script
Creates script used to initialize Lotus. This script asks the user for
a number of inputs, including a DigitalOcean API authorization token and
the necessary variables to bootstrap the Lotus command server.

In its current form it accepts only a limited number of arguments and
is specific to DigitalOcean. In future releases there will either be
multiple initialization scripts (one for each API) or preferably a
single script that uses abstract methods depending on the user's API
selection.