Reconfigures server cloud-config (cron/logs/etc.)
Applies numerous fixes and changes to the command server cloud-config
file. Primarily, shell scripts (such as bootstrap.sh) run by /etc/cron.d
do not include the user's PATH variable -- this must be set manually,
either in the script after the shebang or in the cron.d job. By placing
it in the cron job file, the bootstrap script is unaware of any changes
to PATH and is therefore PATH-agnostic.

All available required packages are now installed by cloud-config, and
file locations have be standardized around the Lotus repo cloned to
/root/lotus and fixes all permissions issues.

/var/log/lotus is now the standard place for Lotus logs.