Reconfigure command server cloud-config runcmd
Due to extensive issues with the write_files module to create a cron.d
job, the command server now uses a series of runcmd (shell commands) to
both create and set permissions for the bootstrap cron.d job.

Also uses a series of runcmds to download the proper version of Lotus.